Implement interruptive multiblock read
Reason/Description: Because right now multiblock read is possible, but only if the destination buffer is large enough. Now it is possible to start a block read, reading a block, reading another block, ... and then stopping the read. This reduces a lot of overhead
Files: lib.rs, mod.rs, and volume.rs